refactor(test): 移动VerifyObject、VerifyFailedField位置
