+++ src/test/java/cn/sticki/validator/spel/bean/SpelNotBlankTestBean.java
@@ -1,9 +1,9 @@
 package cn.sticki.validator.spel.bean;
 
 import cn.sticki.validator.spel.SpelValid;
-import cn.sticki.validator.spel.VerifyFailedField;
-import cn.sticki.validator.spel.VerifyObject;
 import cn.sticki.validator.spel.constrain.SpelNotBlank;
+import cn.sticki.validator.spel.util.VerifyFailedField;
+import cn.sticki.validator.spel.util.VerifyObject;
 import lombok.Data;
 
 import java.util.ArrayList;
