+++ src/test/java/cn/sticki/validator/spel/bean/SpelValidTestBean.java
@@ -1,10 +1,10 @@
 package cn.sticki.validator.spel.bean;
 
 import cn.sticki.validator.spel.SpelValid;
-import cn.sticki.validator.spel.VerifyFailedField;
-import cn.sticki.validator.spel.VerifyObject;
 import cn.sticki.validator.spel.constrain.SpelNotNull;
 import cn.sticki.validator.spel.util.ID;
+import cn.sticki.validator.spel.util.VerifyFailedField;
+import cn.sticki.validator.spel.util.VerifyObject;
 import lombok.Builder;
 import lombok.Data;
